+++ mybatis-plus-support/src/main/java/com/baomidou/mybatisplus/mapper/AutoSqlInjector.java
@@ -17,12 +17,21 @@ package com.baomidou.mybatisplus.mapper;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
+import java.lang.reflect.TypeVariable;
+import java.lang.reflect.WildcardType;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import com.baomidou.mybatisplus.toolkit.ArrayUtils;
+import com.baomidou.mybatisplus.toolkit.CollectionUtils;
+import com.baomidou.mybatisplus.toolkit.GlobalConfigUtils;
+import com.baomidou.mybatisplus.toolkit.PluginUtils;
+import com.baomidou.mybatisplus.toolkit.SqlReservedWords;
+import com.baomidou.mybatisplus.toolkit.StringUtils;
+import com.baomidou.mybatisplus.toolkit.TableInfoHelper;
 import org.apache.ibatis.builder.MapperBuilderAssistant;
 import org.apache.ibatis.executor.keygen.Jdbc3KeyGenerator;
 import org.apache.ibatis.executor.keygen.KeyGenerator;
@@ -46,11 +55,6 @@ import com.baomidou.mybatisplus.enums.FieldFill;
 import com.baomidou.mybatisplus.enums.FieldStrategy;
 import com.baomidou.mybatisplus.enums.IdType;
 import com.baomidou.mybatisplus.enums.SqlMethod;
-import com.baomidou.mybatisplus.toolkit.CollectionUtils;
-import com.baomidou.mybatisplus.toolkit.GlobalConfigUtils;
-import com.baomidou.mybatisplus.toolkit.SqlReservedWords;
-import com.baomidou.mybatisplus.toolkit.StringUtils;
-import com.baomidou.mybatisplus.toolkit.TableInfoHelper;
 
 /**
  * <p>
@@ -69,7 +73,9 @@ public class AutoSqlInjector implements ISqlInjector {
     protected MapperBuilderAssistant builderAssistant;
 
     /**
-     * CRUD注入后给予标识 注入过后不再注入
+     * <p>
+     * CRUD 注入后给予标识 注入过后不再注入
+     * </p>
      *
      * @param builderAssistant
      * @param mapperClass
@@ -92,7 +98,7 @@ public class AutoSqlInjector implements ISqlInjector {
         this.configuration = builderAssistant.getConfiguration();
         this.builderAssistant = builderAssistant;
         this.languageDriver = configuration.getDefaultScriptingLanguageInstance();
-        /*
+        /**
          * 驼峰设置 PLUS 配置 > 原始配置
 		 */
         GlobalConfiguration globalCache = this.getGlobalConfig();
@@ -100,7 +106,13 @@ public class AutoSqlInjector implements ISqlInjector {
             globalCache.setDbColumnUnderline(configuration.isMapUnderscoreToCamelCase());
         }
         Class<?> modelClass = extractModelClass(mapperClass);
-        if (modelClass != null) {
+        if (null != modelClass) {
+            /**
+             * 初始化 SQL 解析
+             */
+            if (globalCache.isSqlParserCache()) {
+                PluginUtils.initSqlParserInfoCache(mapperClass);
+            }
             TableInfo table = TableInfoHelper.initTableInfo(builderAssistant, modelClass);
             injectSql(builderAssistant, mapperClass, modelClass, table);
         }
@@ -121,13 +133,13 @@ public class AutoSqlInjector implements ISqlInjector {
          * #148 表信息包含主键，注入主键相关方法
          */
         if (StringUtils.isNotEmpty(table.getKeyProperty())) {
-            /* 删除 */
+            /** 删除 */
             this.injectDeleteByIdSql(false, mapperClass, modelClass, table);
             this.injectDeleteByIdSql(true, mapperClass, modelClass, table);
-            /* 修改 */
+            /** 修改 */
             this.injectUpdateByIdSql(true, mapperClass, modelClass, table);
             this.injectUpdateByIdSql(false, mapperClass, modelClass, table);
-            /* 查询 */
+            /** 查询 */
             this.injectSelectByIdSql(false, mapperClass, modelClass, table);
             this.injectSelectByIdSql(true, mapperClass, modelClass, table);
         } else {
@@ -138,15 +150,15 @@ public class AutoSqlInjector implements ISqlInjector {
         /**
          * 正常注入无需主键方法
          */
-        /* 插入 */
+        /** 插入 */
         this.injectInsertOneSql(true, mapperClass, modelClass, table);
         this.injectInsertOneSql(false, mapperClass, modelClass, table);
-        /* 删除 */
+        /** 删除 */
         this.injectDeleteSql(mapperClass, modelClass, table);
         this.injectDeleteByMapSql(mapperClass, table);
-        /* 修改 */
+        /** 修改 */
         this.injectUpdateSql(mapperClass, modelClass, table);
-        /* 查询 */
+        /** 查询 */
         this.injectSelectByMapSql(mapperClass, modelClass, table);
         this.injectSelectOneSql(mapperClass, modelClass, table);
         this.injectSelectCountSql(mapperClass, modelClass, table);
@@ -155,7 +167,7 @@ public class AutoSqlInjector implements ISqlInjector {
         this.injectSelectMapsSql(SqlMethod.SELECT_MAPS, mapperClass, modelClass, table);
         this.injectSelectMapsSql(SqlMethod.SELECT_MAPS_PAGE, mapperClass, modelClass, table);
         this.injectSelectObjsSql(SqlMethod.SELECT_OBJS, mapperClass, modelClass, table);
-        /* 自定义方法 */
+        /** 自定义方法 */
         this.inject(configuration, builderAssistant, mapperClass, modelClass, table);
     }
 
@@ -168,26 +180,32 @@ public class AutoSqlInjector implements ISqlInjector {
     }
 
     /**
-     * 避免扫描到BaseMapper
+     * 提取泛型模型,多泛型的时候请将泛型T放在第一位
      *
      * @param mapperClass
      * @return
      */
     protected Class<?> extractModelClass(Class<?> mapperClass) {
-        if (mapperClass == BaseMapper.class) {
-            logger.warn(" Current Class is BaseMapper ");
-            return null;
-        } else {
-            Type[] types = mapperClass.getGenericInterfaces();
-            ParameterizedType target = null;
-            for (Type type : types) {
-                if (type instanceof ParameterizedType && BaseMapper.class.isAssignableFrom(mapperClass)) {
-                    target = (ParameterizedType) type;
-                    break;
+        Type[] types = mapperClass.getGenericInterfaces();
+        ParameterizedType target = null;
+        for (Type type : types) {
+            if (type instanceof ParameterizedType) {
+                Type[] typeArray = ((ParameterizedType) type).getActualTypeArguments();
+                if(ArrayUtils.isNotEmpty(typeArray)){
+                    for (Type t:typeArray) {
+                        if(t instanceof TypeVariable || t instanceof WildcardType){
+                            target = null;
+                            break;
+                        }else {
+                            target = (ParameterizedType) type;
+                            break;
+                        }
+                    }
                 }
+                break;
             }
-            return target == null ? null : (Class<?>) target.getActualTypeArguments()[0];
         }
+        return target == null ? null : (Class<?>) target.getActualTypeArguments()[0];
     }
 
     /**
@@ -220,7 +238,7 @@ public class AutoSqlInjector implements ISqlInjector {
         // 表包含主键处理逻辑,如果不包含主键当普通字段处理
         if (StringUtils.isNotEmpty(table.getKeyProperty())) {
             if (table.getIdType() == IdType.AUTO) {
-                /* 自增主键 */
+                /** 自增主键 */
                 keyGenerator = new Jdbc3KeyGenerator();
                 keyProperty = table.getKeyProperty();
                 keyColumn = table.getKeyColumn();
@@ -232,7 +250,7 @@ public class AutoSqlInjector implements ISqlInjector {
                     fieldBuilder.append(table.getKeyColumn()).append(",");
                     placeholderBuilder.append("#{").append(table.getKeyProperty()).append("},");
                 } else {
-                    /* 用户输入自定义ID */
+                    /** 用户输入自定义ID */
                     fieldBuilder.append(table.getKeyColumn()).append(",");
                     // 正常自定义主键策略
                     placeholderBuilder.append("#{").append(table.getKeyProperty()).append("},");
@@ -494,6 +512,15 @@ public class AutoSqlInjector implements ISqlInjector {
 
     /**
      * <p>
+     * Sql 运算条件
+     * </p>
+     */
+    protected String sqlCondition(String condition, String column, String property) {
+        return String.format(condition, column, property);
+    }
+
+    /**
+     * <p>
      * EntityWrapper方式获取select where
      * </p>
      *
@@ -513,7 +540,8 @@ public class AutoSqlInjector implements ISqlInjector {
         List<TableFieldInfo> fieldList = table.getFieldList();
         for (TableFieldInfo fieldInfo : fieldList) {
             where.append(convertIfTag(fieldInfo, "ew.entity.", false));
-            where.append(" AND ").append(fieldInfo.getColumn()).append("=#{ew.entity.").append(fieldInfo.getEl()).append("}");
+            where.append(" AND ").append(this.sqlCondition(fieldInfo.getCondition(),
+                fieldInfo.getColumn(), "ew.entity." + fieldInfo.getEl()));
             where.append(convertIfTag(fieldInfo, true));
         }
         where.append("\n</if>");
@@ -546,13 +574,18 @@ public class AutoSqlInjector implements ISqlInjector {
             ifTag = !(FieldFill.UPDATE == fieldInfo.getFieldFill()
                     || FieldFill.INSERT_UPDATE == fieldInfo.getFieldFill());
             if (selective && ifTag) {
-                set.append(convertIfTag(true, fieldInfo, prefix, false));
-                set.append(fieldInfo.getColumn()).append("=#{");
-                if (null != prefix) {
-                    set.append(prefix);
+                if (StringUtils.isNotEmpty(fieldInfo.getUpdate())) {
+                    set.append(fieldInfo.getColumn()).append("=");
+                    set.append(String.format(fieldInfo.getUpdate(), fieldInfo.getColumn())).append(",");
+                } else {
+                    set.append(convertIfTag(true, fieldInfo, prefix, false));
+                    set.append(fieldInfo.getColumn()).append("=#{");
+                    if (null != prefix) {
+                        set.append(prefix);
+                    }
+                    set.append(fieldInfo.getEl()).append("},");
+                    set.append(convertIfTag(true, fieldInfo, null, true));
                 }
-                set.append(fieldInfo.getEl()).append("},");
-                set.append(convertIfTag(true, fieldInfo, null, true));
             } else if (FieldFill.INSERT != fieldInfo.getFieldFill()) {
                 // 排除填充注解字段
                 set.append(fieldInfo.getColumn()).append("=#{");
@@ -716,7 +749,8 @@ public class AutoSqlInjector implements ISqlInjector {
         List<TableFieldInfo> fieldList = table.getFieldList();
         for (TableFieldInfo fieldInfo : fieldList) {
             where.append(convertIfTag(fieldInfo, "ew.", false));
-            where.append(" AND ").append(fieldInfo.getColumn()).append("=#{ew.").append(fieldInfo.getEl()).append("}");
+            where.append(" AND ").append(this.sqlCondition(fieldInfo.getCondition(),
+                fieldInfo.getColumn(), "ew." + fieldInfo.getEl()));
             where.append(convertIfTag(fieldInfo, true));
         }
         where.append("\n</where>");
@@ -754,7 +788,7 @@ public class AutoSqlInjector implements ISqlInjector {
      * @return
      */
     protected String convertIfTag(boolean ignored, TableFieldInfo fieldInfo, String prefix, boolean close) {
-        /* 忽略策略 */
+        /** 忽略策略 */
         FieldStrategy fieldStrategy = fieldInfo.getFieldStrategy();
         if (fieldStrategy == FieldStrategy.IGNORED) {
             if (ignored) {
@@ -769,7 +803,7 @@ public class AutoSqlInjector implements ISqlInjector {
             return "</if>";
         }
 
-		/* 前缀处理 */
+		/** 前缀处理 */
         String property = fieldInfo.getProperty();
         Class propertyType = fieldInfo.getPropertyType();
         property = StringUtils.removeIsPrefixIfBoolean(property, propertyType);
@@ -801,7 +835,7 @@ public class AutoSqlInjector implements ISqlInjector {
         return convertIfTag(fieldInfo, null, close);
     }
 
-    /*
+    /**
      * 查询
      */
     public MappedStatement addSelectMappedStatement(Class<?> mapperClass, String id, SqlSource sqlSource, Class<?> resultType,
@@ -809,18 +843,18 @@ public class AutoSqlInjector implements ISqlInjector {
         if (null != table) {
             String resultMap = table.getResultMap();
             if (null != resultMap) {
-                /* 返回 resultMap 映射结果集 */
+                /** 返回 resultMap 映射结果集 */
                 return this.addMappedStatement(mapperClass, id, sqlSource, SqlCommandType.SELECT, null, resultMap, null,
                         new NoKeyGenerator(), null, null);
             }
         }
 
-		/* 普通查询 */
+		/** 普通查询 */
         return this.addMappedStatement(mapperClass, id, sqlSource, SqlCommandType.SELECT, null, null, resultType,
                 new NoKeyGenerator(), null, null);
     }
 
-    /*
+    /**
      * 插入
      */
     public MappedStatement addInsertMappedStatement(Class<?> mapperClass, Class<?> modelClass, String id, SqlSource sqlSource,
@@ -829,7 +863,7 @@ public class AutoSqlInjector implements ISqlInjector {
                 keyGenerator, keyProperty, keyColumn);
     }
 
-    /*
+    /**
      * 删除
      */
     public MappedStatement addDeleteMappedStatement(Class<?> mapperClass, String id, SqlSource sqlSource) {
@@ -837,7 +871,7 @@ public class AutoSqlInjector implements ISqlInjector {
                 new NoKeyGenerator(), null, null);
     }
 
-    /*
+    /**
      * 更新
      */
     public MappedStatement addUpdateMappedStatement(Class<?> mapperClass, Class<?> modelClass, String id, SqlSource sqlSource) {
@@ -849,12 +883,12 @@ public class AutoSqlInjector implements ISqlInjector {
                                               SqlCommandType sqlCommandType, Class<?> parameterClass, String resultMap, Class<?> resultType,
                                               KeyGenerator keyGenerator, String keyProperty, String keyColumn) {
         String statementName = mapperClass.getName() + "." + id;
-        if (configuration.hasStatement(statementName)) {
+        if (hasMappedStatement(statementName)) {
             System.err.println("{" + statementName
                     + "} Has been loaded by XML or SqlProvider, ignoring the injection of the SQL.");
             return null;
         }
-		/* 缓存逻辑处理 */
+		/** 缓存逻辑处理 */
         boolean isSelect = false;
         if (sqlCommandType == SqlCommandType.SELECT) {
             isSelect = true;